+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased changes
 
+- Fix: change query and header for servant-foreign-0.11
+    - see: https://github.com/haskell-servant/servant/pull/843
+
 ## 0.1.0
 
 - Initial release !